requirements.txt
matplotlib==3.4.3
pandas==1.3.3
Pillow==8.3.2
tensorflow==2.6.0
scipy==1.7.1
h5py==3.1.0
requests==2.26.0
seaborn==0.11.2
tensorflow-gpu~=2.7.0
numpy~=1.22.2